Added metadata to embeds
Embeds used for the poll and help commands now get metadata via adding an url to the author.
This url is a valid url string with the metadata added in the form of parameters.
Additionaly, the poll metadata includes support for different poll types via an additional integer indicating the type.

The old poll style is currently still supported, but will soon be deprecated